requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
urllib3==2.0.4
pytest==7.4.0
selenium==4.11.2

